requirements.txt
yt-dlp>=2023.11.16
pywebview>=4.4.1